pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.33.0",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "streamlit>=1.49.1",
]
